+++ client/advertising/advertising.html
@@ -22,7 +22,7 @@
           進行中的廣告
         </div>
         {{#if currentUser}}
-          <div class="d-flex flex-wrap justify-content-end ml-auto mt-1">
+          <div class="d-flex flex-wrap justify-content-end ml-auto">
             <button class="btn btn-primary btn-sm" type="button" data-action="buyAdvertising">
               <i class="fa fa-plus" aria-hidden="true"></i>
               購買廣告
