+++ client/announcement/announcement.html
@@ -6,7 +6,7 @@
           系統公告
         </div>
         {{#if currentUser.profile.isAdmin}}
-          <div class="d-flex flex-wrap justify-content-end ml-auto mt-1">
+          <div class="d-flex flex-wrap justify-content-end ml-auto">
             <button class="btn btn-warning btn-sm" type="button" data-action="editAnnouncement">
               <i class="fa fa-pencil-square-o" aria-hidden="true"></i>
               修改系統公告
