+++ client/layout/bigPicturePreviewModal.html
@@ -1,0 +1,16 @@
+<template name="bigPicturePreviewModal">
+  <div class="{{modalClass}}" tabindex="-1" role="dialog" style="background-color: rgba(0,0,0,0.4)">
+    <div class="modal-dialog" role="document">
+      <div class="modal-content border-0 mx-0 my-auto"
+        style="background: transparent;">
+        <div class="modal-body p-0" style="width: 300px; height: 300px;">
+          <img
+            class="w-100 h-100"
+            style="object-fit: scale-down;"
+            src="{{src}}"
+            />
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
