+++ client/layout/footer.js
@@ -7,7 +7,6 @@ import { dbAdvertising } from '/db/dbAdvertising';
 import { dbVariables } from '/db/dbVariables';
 import { rMainTheme } from '../utils/styles';
 import { shouldStopSubscribe } from '../utils/idle';
-import { externalLinkAlert } from '../utils/externalLinkAlert';
 
 Template.footer.onCreated(function() {
   this.autorun(() => {
@@ -112,19 +111,17 @@ Template.displayAnnouncement.helpers({
   }
 });
 
-const displayAnnouncementEvents = Object.assign({
+Template.displayAnnouncement.events({
   'click .btn'(event) {
     event.preventDefault();
     rIsDisplayAnnouncement.set(false);
   }
-}, externalLinkAlert);
-Template.displayAnnouncement.events(displayAnnouncementEvents);
+});
 
-const displayAdvertisingEvents = Object.assign({
+Template.displayAdvertising.events({
   'click .btn'(event, templateInstance) {
     event.preventDefault();
     const closedAdvertisingIdList = rClosedAdvertisingIdList.get().slice();
     rClosedAdvertisingIdList.set(_.union(closedAdvertisingIdList, templateInstance.data._id));
   }
-}, externalLinkAlert);
-Template.displayAdvertising.events(displayAdvertisingEvents);
+});
