+++ client/layout/layout.html
@@ -40,4 +40,5 @@
   {{#if showAlertDialog}}
     {{> alertDialog}}
   {{/if}}
+  {{> bigPicturePreviewModal}}
 </template>
